Add support for proxying the media server's URLs via `/proxy`
This feature is enabled with `vibin serve --proxy-media-server`.

The use case for proxying is when the UI will have access to the vibin
server but won't have access to the media server device (perhaps
because the UI is being accessed from another network).

Any payloads (REST or WebSocket) sent back to the UI which contain
references to the media server (e.g. album art URLs) will instead have
those references be replaced by `/proxy`. `/proxy` will then act as a
reverse proxy to the media server's endpoints.

The downside to this is that more data is flowing through the vibin
server, which will introduce additional load and potentially impact
performance.