Improve stability of WebSocket connections to connected clients
The system seemed to get into a state where it wouldn't recover from
sending a message to a client which was in an unhappy state (which
may have been related to the client -- a web browser -- having its host
machine go into sleep). The await on the send never seemed to return.
This PR (the guts of which moves the send into a try block) seems to
help, but future investigations into WebSocket handling may be required.