Determine currently-playing local media ids at startup
This ensures that the active album and track ids are known for the
currently-playing media at startup time, rather than only setting those
IDs when the track changes.

This only impacts the period of time between startup and the first
track change, as subsequent track changes trigger a UPnP event which in
turn sets the media IDs.