Improve artist determination in links manager
Extends the artist-determination heuristic in
`_artist_name_from_track_media_info()` to allow for another potential
case. This logic is getting a little unwieldy and should be extracted
at some point. It would also be good to dig into the UPnP specification
to figure out what the real possibilities for artist name are, rather
than going off of guesswork and seen cases.