Move from one TinyDB file to one file per table
This initializes a separate TinyDB file per database table. It's
possible to instead store all tables in one file, but that affects
performance as the file size grows. The downside to multiple files is
that we have a pretty sloppy collection of files, TinyDB instances, and
DB locks. TinyDB should be replaced with another database solution
TinyDB should be replaced with another database solution.