Use `volume_step` instead of `volume_percent`
Fixes a bug where vibinui causes the volume to slowly decrease until it hits a multiple of 10%.

The CXNv2 itself has 30 steps of volume. It appears to round values down when applying a percentage volume, and again when reporting the new volume back via the `volume_percent` field.

Hence, setting the volume to, say, 95% causes the streamer to set its volume to floor(30*95%) = 28, which is reported back as floor(28/30) = 93%.

This, in turn, triggers the UI's volume change detection logic in `KeyboardShortcutsManager`, which bounces the new volume value of 93% back to the streamer.

But the streamer then interprets this requested value of 93% as floor(30*93%) = 27, which gets reported back as floor(27/30) = 90%. Once again, the UI bounces that new volume of 90% back to the streamer, but because floor(30*90%) is still 27, this is now stable.

This server-side fix rounds to the nearest available step, which means that the reported fractional volume can still differ from the requested fractional volume. Under some circumstances the client still bounces that changed volume back, but it remains stable.