Allow the streamer device to be the amplifier if no other amplifiers are discovered
This appears to be the intention of the existing code.

The Streammagic amplifier implementation will be used by default for a Streammagic streamer. Although the UI will report this "amplifier" as present, the UI will only offer volume controls if the streamer is configured with Control Bus or Pre-Amp mode.