Use a streamer's natural `Amplifier` implementation if no others found (#135)
* Don't report power for Streammagic amplifier

The power status of the amplifier is unknown to Streammagic, but is usually synced to the power state of the streamer using the control bus.

This allows the client to avoid spamming the user with redundant notifications when the streamer's power is toggled.

* Allow the streamer device to be the amplifier if no other amplifiers are discovered

This appears to be the intention of the existing code.

The Streammagic amplifier implementation will be used by default for a Streammagic streamer. Although the UI will report this "amplifier" as present, the UI will only offer volume controls if the streamer is configured with Control Bus or Pre-Amp mode.