Add WIP for supporting streaming amplifiers (like the Evo series)
This is an incomplete WIP.

The Cambridge Audio streaming amplifiers don't seem to provide some of
the UPnP services relied on by Vibin -- like UuVolControl and
PlaylistExtension -- so Vibin's dependency on them needs to be removed.
On the upside this will remove the majority of Vibin's awkward-feeling
UPnP service code. On the downside, it may break support for waveform
generation as the ability to get the actual network filename for the
local media being played was provided via one of those UPnP services.

This also includes some tweaks to make the Asset UPnP MediaServer
implementation a little more resilient. That should probably be broken
out into a separate PR later on.